% tinyrox says don't edit this manually, but it can't stop you!
\name{as.character.torch_dtype}
\alias{as.character.torch_dtype}
\title{As.character.torch dtype}
\usage{
\method{as.character}{torch_dtype}(x, ...)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}

\item{...}{Additional arguments passed to methods.}
}
\description{
As.character.torch dtype
}
\examples{
as.character(torch_float32())
}
