% tinyrox says don't edit this manually, but it can't stop you!
\name{autograd_function}
\alias{autograd_function}
\title{Define a custom autograd function}
\usage{
autograd_function(forward, backward)
}
\arguments{
\item{forward}{Forward function.}

\item{backward}{Backward function.}
}
\value{
An autograd function class.
}
\description{
Define a custom autograd function
}
