% tinyrox says don't edit this manually, but it can't stop you!
\name{autograd_set_grad_mode}
\alias{autograd_set_grad_mode}
\title{Autograd set grad mode}
\usage{
autograd_set_grad_mode(enabled)
}
\arguments{
\item{enabled}{Parameter passed to the underlying ATen operator.}
}
\description{
Autograd set grad mode
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
