% tinyrox says don't edit this manually, but it can't stop you!
\name{backends_cudnn_is_available}
\alias{backends_cudnn_is_available}
\title{Backends cudnn is available}
\usage{
backends_cudnn_is_available()
}
\description{
Backends cudnn is available
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
