% tinyrox says don't edit this manually, but it can't stop you!
\name{backends_cudnn_version}
\alias{backends_cudnn_version}
\title{Backends cudnn version}
\usage{
backends_cudnn_version()
}
\value{
A `torch_tensor`.
}
\description{
Backends cudnn version
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
