% tinyrox says don't edit this manually, but it can't stop you!
\name{backends_mkldnn_is_available}
\alias{backends_mkldnn_is_available}
\title{Backends mkldnn is available}
\usage{
backends_mkldnn_is_available()
}
\description{
Backends mkldnn is available
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
