% tinyrox says don't edit this manually, but it can't stop you!
\name{backends_mps_is_available}
\alias{backends_mps_is_available}
\title{Backends mps is available}
\usage{
backends_mps_is_available()
}
\value{
A `torch_tensor`.
}
\description{
Backends mps is available
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
