% tinyrox says don't edit this manually, but it can't stop you!
\name{clear_exec_cache}
\alias{clear_exec_cache}
\title{Clear the Execution Cache}
\usage{
clear_exec_cache()
}
\value{
Integer, number of entries cleared (invisibly)
}
\description{
Removes all cached prepared graphs from the auto-cache used by
\code{execute_optimized()}.
}
\examples{
clear_exec_cache()
}
