% tinyrox says don't edit this manually, but it can't stop you!
\name{clear_gpu_kernel_cache}
\alias{clear_gpu_kernel_cache}
\title{Clear GPU Kernel Cache}
\usage{
clear_gpu_kernel_cache()
}
\value{
Number of entries cleared (invisibly)
}
\description{
Clear GPU Kernel Cache
}
\examples{
clear_gpu_kernel_cache()
}
