% tinyrox says don't edit this manually, but it can't stop you!
\name{clear_kernel_cache}
\alias{clear_kernel_cache}
\title{Clear Kernel Disk Cache}
\usage{
clear_kernel_cache()
}
\value{
Number of cached kernels cleared (invisibly)
}
\description{
Removes all compiled kernel shared libraries from the disk cache
and clears the in-memory registry.
}
