% tinyrox says don't edit this manually, but it can't stop you!
\name{clone_module}
\alias{clone_module}
\title{Clone module}
\usage{
clone_module(module)
}
\arguments{
\item{module}{Parameter passed to the underlying ATen operator.}
}
\value{
A cloned `nn_module`.
}
\description{
Clone module
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
