% tinyrox says don't edit this manually, but it can't stop you!
\name{cuda_current_device}
\alias{cuda_current_device}
\title{Cuda current device}
\usage{
cuda_current_device()
}
\value{
Integer scalar.
}
\description{
Cuda current device
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
