% tinyrox says don't edit this manually, but it can't stop you!
\name{cuda_get_device_capability}
\alias{cuda_get_device_capability}
\title{Cuda get device capability}
\usage{
cuda_get_device_capability(device = 0L)
}
\arguments{
\item{device}{Parameter passed to the underlying ATen operator.}
}
\description{
Cuda get device capability
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
