% tinyrox says don't edit this manually, but it can't stop you!
\name{cuda_get_rng_state}
\alias{cuda_get_rng_state}
\title{Cuda get rng state}
\usage{
cuda_get_rng_state(device = 0L)
}
\arguments{
\item{device}{Parameter passed to the underlying ATen operator.}
}
\description{
Cuda get rng state
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
