% tinyrox says don't edit this manually, but it can't stop you!
\name{cuda_runtime_version}
\alias{cuda_runtime_version}
\title{Cuda runtime version}
\usage{
cuda_runtime_version()
}
\description{
Cuda runtime version
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
