% tinyrox says don't edit this manually, but it can't stop you!
\name{.find_modules_runtime}
\alias{.find_modules_runtime}
\title{Find nn_modules at Runtime in an Installed Package}
\usage{
.find_modules_runtime(pkg_name)
}
\description{
Find nn_modules at Runtime in an Installed Package
}
\keyword{internal}
