% tinyrox says don't edit this manually, but it can't stop you!
\name{exec_cache_stats}
\alias{exec_cache_stats}
\title{Execution Cache Statistics}
\usage{
exec_cache_stats()
}
\value{
A list with \code{n_cached} (number of cached prepared graphs)
}
\description{
Returns information about the auto-cache used by
\code{execute_optimized()}.
}
\examples{
exec_cache_stats()
}
