% tinyrox says don't edit this manually, but it can't stop you!
\name{gpu_kernel_cache_stats}
\alias{gpu_kernel_cache_stats}
\title{GPU kernel cache statistics}
\usage{
gpu_kernel_cache_stats()
}
\value{
List with n_cached and kernel_names.
}
\description{
GPU kernel cache statistics
}
\keyword{internal}
