% tinyrox says don't edit this manually, but it can't stop you!
\name{install_torch}
\alias{install_torch}
\title{Install torch}
\usage{
install_torch(...)
}
\arguments{
\item{...}{Additional arguments passed to methods.}
}
\description{
Install torch
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
