% tinyrox says don't edit this manually, but it can't stop you!
\name{is_dataloader}
\alias{is_dataloader}
\title{Is dataloader}
\usage{
is_dataloader(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\description{
Is dataloader
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
