% tinyrox says don't edit this manually, but it can't stop you!
\name{is_nn_buffer}
\alias{is_nn_buffer}
\title{Is nn buffer}
\usage{
is_nn_buffer(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\value{
Logical scalar.
}
\description{
Is nn buffer
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
