% tinyrox says don't edit this manually, but it can't stop you!
\name{is_nn_parameter}
\alias{is_nn_parameter}
\title{Is nn parameter}
\usage{
is_nn_parameter(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\description{
Is nn parameter
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
