% tinyrox says don't edit this manually, but it can't stop you!
\name{is_optimizer}
\alias{is_optimizer}
\title{Is optimizer}
\usage{
is_optimizer(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\description{
Is optimizer
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
