% tinyrox says don't edit this manually, but it can't stop you!
\name{is_torch_device}
\alias{is_torch_device}
\title{Is torch device}
\usage{
is_torch_device(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\description{
Is torch device
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
