% tinyrox says don't edit this manually, but it can't stop you!
\name{is_torch_dtype}
\alias{is_torch_dtype}
\title{Is torch dtype}
\usage{
is_torch_dtype(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\value{
Logical scalar.
}
\description{
Is torch dtype
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
