% tinyrox says don't edit this manually, but it can't stop you!
\name{is_torch_qscheme}
\alias{is_torch_qscheme}
\title{Is torch qscheme}
\usage{
is_torch_qscheme(x)
}
\arguments{
\item{x}{Parameter passed to the underlying ATen operator.}
}
\description{
Is torch qscheme
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
