% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{iterable_dataset}
\alias{iterable_dataset}
\title{Iterable dataset (stub)}
\format{
An object of class \code{function} of length 1.
}
\usage{
iterable_dataset
}
\description{
Iterable dataset (stub)
}
\examples{
\donttest{
if (torch_is_installed()) {
# See ?dataset for usage; iterable_dataset is currently an alias
}
}
}
\keyword{datasets}
