% tinyrox says don't edit this manually, but it can't stop you!
\name{load_state_dict}
\alias{load_state_dict}
\title{Load state dict}
\usage{
load_state_dict(module, state_dict)
}
\arguments{
\item{module}{Parameter passed to the underlying ATen operator.}

\item{state_dict}{Parameter passed to the underlying ATen operator.}
}
\description{
Load state dict
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
