% tinyrox says don't edit this manually, but it can't stop you!
\name{loop}
\alias{loop}
\title{Loop over a dataloader (convenience)}
\usage{
loop(dl, fn)
}
\arguments{
\item{dl}{A dataloader.}

\item{fn}{Function to call with each batch.}
}
\value{
Invisible NULL.
}
\description{
Loop over a dataloader (convenience)
}
