% tinyrox says don't edit this manually, but it can't stop you!
\name{lr_lambda}
\alias{lr_lambda}
\title{Lr lambda}
\usage{
lr_lambda(optimizer, lr_lambda)
}
\arguments{
\item{optimizer}{Parameter passed to the underlying ATen operator.}

\item{lr_lambda}{Parameter passed to the underlying ATen operator.}
}
\value{
A learning rate scheduler.
}
\description{
Lr lambda
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
