% tinyrox says don't edit this manually, but it can't stop you!
\name{lr_one_cycle}
\alias{lr_one_cycle}
\title{Lr one cycle}
\usage{
lr_one_cycle(optimizer, max_lr, total_steps)
}
\arguments{
\item{optimizer}{Parameter passed to the underlying ATen operator.}

\item{max_lr}{Parameter passed to the underlying ATen operator.}

\item{total_steps}{Parameter passed to the underlying ATen operator.}
}
\description{
Lr one cycle
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
