% tinyrox says don't edit this manually, but it can't stop you!
\name{lr_scheduler}
\alias{lr_scheduler}
\title{Lr scheduler}
\usage{
lr_scheduler(optimizer, ...)
}
\arguments{
\item{optimizer}{Parameter passed to the underlying ATen operator.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A learning rate scheduler.
}
\description{
Lr scheduler
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
