% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_adaptive_avg_pool1d}
\alias{nn_adaptive_avg_pool1d}
\title{Nn adaptive avg pool1d}
\usage{
nn_adaptive_avg_pool1d(output_size)
}
\arguments{
\item{output_size}{Parameter passed to the underlying ATen operator.}
}
\value{
An `nn_module` instance.
}
\description{
Nn adaptive avg pool1d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
