% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_adaptive_max_pool2d}
\alias{nn_adaptive_max_pool2d}
\title{Nn adaptive max pool2d}
\usage{
nn_adaptive_max_pool2d(output_size)
}
\arguments{
\item{output_size}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn adaptive max pool2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
