% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_avg_pool2d}
\alias{nn_avg_pool2d}
\title{Nn avg pool2d}
\usage{
nn_avg_pool2d(kernel_size, stride = NULL, padding = 0L)
}
\arguments{
\item{kernel_size}{Integer vector kernel size.}

\item{stride}{Integer vector stride.}

\item{padding}{Integer vector padding.}
}
\description{
Nn avg pool2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
