% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_bce_with_logits_loss}
\alias{nn_bce_with_logits_loss}
\title{Nn bce with logits loss}
\usage{
nn_bce_with_logits_loss(weight = NULL, pos_weight = NULL, reduction = 1L)
}
\arguments{
\item{weight}{A `torch_tensor` of weights.}

\item{pos_weight}{Parameter passed to the underlying ATen operator.}

\item{reduction}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn bce with logits loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
