% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_celu}
\alias{nn_celu}
\title{Nn celu}
\usage{
nn_celu(alpha = 1)
}
\arguments{
\item{alpha}{Numeric scalar multiplier.}
}
\description{
Nn celu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
