% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_dropout2d}
\alias{nn_dropout2d}
\title{Nn dropout2d}
\usage{
nn_dropout2d(p = 0.5)
}
\arguments{
\item{p}{Numeric norm exponent.}
}
\value{
An `nn_module` instance.
}
\description{
Nn dropout2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
