% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_dropout3d}
\alias{nn_dropout3d}
\title{Nn dropout3d}
\usage{
nn_dropout3d(p = 0.5)
}
\arguments{
\item{p}{Numeric norm exponent.}
}
\description{
Nn dropout3d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
