% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_flatten}
\alias{nn_flatten}
\title{Nn flatten}
\usage{
nn_flatten(start_dim = 2L, end_dim = -1L)
}
\arguments{
\item{start_dim}{Integer starting dimension.}

\item{end_dim}{Integer ending dimension.}
}
\value{
An `nn_module` instance.
}
\description{
Nn flatten
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
