% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_hardshrink}
\alias{nn_hardshrink}
\title{Nn hardshrink}
\usage{
nn_hardshrink(lambd = 0.5)
}
\arguments{
\item{lambd}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn hardshrink
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
