% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_hardsigmoid}
\alias{nn_hardsigmoid}
\title{Nn hardsigmoid}
\usage{
nn_hardsigmoid()
}
\value{
An `nn_module` instance.
}
\description{
Nn hardsigmoid
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
