% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_hardswish}
\alias{nn_hardswish}
\title{Nn hardswish}
\usage{
nn_hardswish()
}
\description{
Nn hardswish
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
