% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_calculate_gain}
\alias{nn_init_calculate_gain}
\title{Nn init calculate gain}
\usage{
nn_init_calculate_gain(nonlinearity, param = NULL)
}
\arguments{
\item{nonlinearity}{Parameter passed to the underlying ATen operator.}

\item{param}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn init calculate gain
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
