% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_constant_}
\alias{nn_init_constant_}
\title{Nn init constant}
\usage{
nn_init_constant_(tensor, val)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}

\item{val}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nn init constant
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
