% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_dirac_}
\alias{nn_init_dirac_}
\title{Dirac initialization}
\usage{
nn_init_dirac_(tensor, groups = 1L)
}
\arguments{
\item{tensor}{A 3D, 4D, or 5D tensor.}

\item{groups}{Number of groups. Default 1.}
}
\value{
tensor (modified in place).
}
\description{
Dirac initialization
}
