% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_eye_}
\alias{nn_init_eye_}
\title{Nn init eye}
\usage{
nn_init_eye_(tensor)
}
\arguments{
\item{tensor}{A `torch_tensor` to be filled in-place with an identity matrix.}
}
\value{
A `torch_tensor`.
}
\description{
Nn init eye
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
