% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_kaiming_uniform_}
\alias{nn_init_kaiming_uniform_}
\title{Nn init kaiming uniform}
\usage{
nn_init_kaiming_uniform_(tensor, a = 0, mode = "fan_in",
                         nonlinearity = "leaky_relu")
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}

\item{a}{Parameter passed to the underlying ATen operator.}

\item{mode}{Parameter passed to the underlying ATen operator.}

\item{nonlinearity}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nn init kaiming uniform
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
