% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_normal_}
\alias{nn_init_normal_}
\title{Nn init normal}
\usage{
nn_init_normal_(tensor, mean = 0, std = 1)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}

\item{mean}{Parameter passed to the underlying ATen operator.}

\item{std}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nn init normal
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
