% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_ones_}
\alias{nn_init_ones_}
\title{Nn init ones}
\usage{
nn_init_ones_(tensor)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn init ones
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
