% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_trunc_normal_}
\alias{nn_init_trunc_normal_}
\title{Nn init trunc normal}
\usage{
nn_init_trunc_normal_(tensor, mean = 0, std = 1, a = -2, b = 2)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}

\item{mean}{Parameter passed to the underlying ATen operator.}

\item{std}{Parameter passed to the underlying ATen operator.}

\item{a}{Parameter passed to the underlying ATen operator.}

\item{b}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn init trunc normal
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
