% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_uniform_}
\alias{nn_init_uniform_}
\title{Nn init uniform}
\usage{
nn_init_uniform_(tensor, a = 0, b = 1)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}

\item{a}{Parameter passed to the underlying ATen operator.}

\item{b}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn init uniform
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
