% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_xavier_uniform_}
\alias{nn_init_xavier_uniform_}
\title{Nn init xavier uniform}
\usage{
nn_init_xavier_uniform_(tensor, gain = 1)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}

\item{gain}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nn init xavier uniform
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
