% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_init_zeros_}
\alias{nn_init_zeros_}
\title{Nn init zeros}
\usage{
nn_init_zeros_(tensor)
}
\arguments{
\item{tensor}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn init zeros
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
