% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_kl_div_loss}
\alias{nn_kl_div_loss}
\title{Nn kl div loss}
\usage{
nn_kl_div_loss(reduction = 1L, log_target = FALSE)
}
\arguments{
\item{reduction}{Parameter passed to the underlying ATen operator.}

\item{log_target}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn kl div loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
