% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_l1_loss}
\alias{nn_l1_loss}
\title{Nn l1 loss}
\usage{
nn_l1_loss(reduction = 1L)
}
\arguments{
\item{reduction}{Reduction string: `"none"`, `"mean"`, or `"sum"`.}
}
\description{
Nn l1 loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
