% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_leaky_relu}
\alias{nn_leaky_relu}
\title{Nn leaky relu}
\usage{
nn_leaky_relu(negative_slope = 0.01)
}
\arguments{
\item{negative_slope}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn leaky relu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
