% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_margin_ranking_loss}
\alias{nn_margin_ranking_loss}
\title{Nn margin ranking loss}
\usage{
nn_margin_ranking_loss(margin = 0, reduction = 1L)
}
\arguments{
\item{margin}{Parameter passed to the underlying ATen operator.}

\item{reduction}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn margin ranking loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
