% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_max_pool2d}
\alias{nn_max_pool2d}
\title{Nn max pool2d}
\usage{
nn_max_pool2d(kernel_size, stride = NULL, padding = 0L, dilation = 1L)
}
\arguments{
\item{kernel_size}{Integer vector kernel size.}

\item{stride}{Integer vector stride.}

\item{padding}{Integer vector padding.}

\item{dilation}{Integer vector dilation.}
}
\description{
Nn max pool2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
