% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_module_list}
\alias{nn_module_list}
\title{List container for nn_modules}
\usage{
nn_module_list(modules = list())
}
\arguments{
\item{modules}{A list of nn_module instances.}
}
\value{
An nn_module containing the sub-modules.
}
\description{
List container for nn_modules
}
