% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_mse_loss}
\alias{nn_mse_loss}
\title{Nn mse loss}
\usage{
nn_mse_loss(reduction = 1L)
}
\arguments{
\item{reduction}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn mse loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
