% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_prelu}
\alias{nn_prelu}
\title{Nn prelu}
\usage{
nn_prelu(num_parameters = 1L)
}
\arguments{
\item{num_parameters}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn prelu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
