% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_prune_head}
\alias{nn_prune_head}
\title{Nn prune head}
\usage{
nn_prune_head(module, ...)
}
\arguments{
\item{module}{Parameter passed to the underlying ATen operator.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Nn prune head
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
