% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_relu6}
\alias{nn_relu6}
\title{Nn relu6}
\usage{
nn_relu6()
}
\value{
An `nn_module` instance.
}
\description{
Nn relu6
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
