% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_rrelu}
\alias{nn_rrelu}
\title{Nn rrelu}
\usage{
nn_rrelu(lower = 0.125, upper = 0.333)
}
\arguments{
\item{lower}{Parameter passed to the underlying ATen operator.}

\item{upper}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn rrelu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
