% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_selu}
\alias{nn_selu}
\title{Nn selu}
\usage{
nn_selu()
}
\value{
An `nn_module` instance.
}
\description{
Nn selu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
