% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_smooth_l1_loss}
\alias{nn_smooth_l1_loss}
\title{Nn smooth l1 loss}
\usage{
nn_smooth_l1_loss(reduction = 1L, beta = 1)
}
\arguments{
\item{reduction}{Reduction string: `"none"`, `"mean"`, or `"sum"`.}

\item{beta}{Numeric scalar multiplier.}
}
\description{
Nn smooth l1 loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
