% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_softmax2d}
\alias{nn_softmax2d}
\title{Nn softmax2d}
\usage{
nn_softmax2d()
}
\description{
Nn softmax2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
