% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_softmin}
\alias{nn_softmin}
\title{Nn softmin}
\usage{
nn_softmin(dim = -1L)
}
\arguments{
\item{dim}{Integer dimension.}
}
\value{
An `nn_module` instance.
}
\description{
Nn softmin
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
