% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_softplus}
\alias{nn_softplus}
\title{Nn softplus}
\usage{
nn_softplus(beta = 1, threshold = 20)
}
\arguments{
\item{beta}{Numeric scalar multiplier.}

\item{threshold}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn softplus
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
