% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_softshrink}
\alias{nn_softshrink}
\title{Nn softshrink}
\usage{
nn_softshrink(lambd = 0.5)
}
\arguments{
\item{lambd}{Parameter passed to the underlying ATen operator.}
}
\value{
An `nn_module` instance.
}
\description{
Nn softshrink
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
