% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_softsign}
\alias{nn_softsign}
\title{Nn softsign}
\usage{
nn_softsign()
}
\value{
An `nn_module` instance.
}
\description{
Nn softsign
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
