% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_threshold}
\alias{nn_threshold}
\title{Nn threshold}
\usage{
nn_threshold(threshold, value)
}
\arguments{
\item{threshold}{Parameter passed to the underlying ATen operator.}

\item{value}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn threshold
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
