% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_unflatten}
\alias{nn_unflatten}
\title{Nn unflatten}
\usage{
nn_unflatten(dim, unflattened_size)
}
\arguments{
\item{dim}{Integer dimension.}

\item{unflattened_size}{Parameter passed to the underlying ATen operator.}
}
\description{
Nn unflatten
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
