% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_upsample}
\alias{nn_upsample}
\title{Upsample module}
\usage{
nn_upsample(size = NULL, scale_factor = NULL, mode = "nearest")
}
\arguments{
\item{size}{Target size (integer or integer vector).}

\item{scale_factor}{Scale factor (double or double vector).}

\item{mode}{Upsampling mode. Default "nearest".}
}
\value{
An nn_module instance.
}
\description{
Upsample module
}
