% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_utils_weight_norm}
\alias{nn_utils_weight_norm}
\title{Weight norm (stub)}
\usage{
nn_utils_weight_norm(module, name = "weight", dim = 0L)
}
\arguments{
\item{module}{An nn_module.}

\item{name}{Parameter name. Default "weight".}

\item{dim}{Dimension. Default 0.}
}
\value{
The module (unmodified stub).
}
\description{
Weight norm (stub)
}
