% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_adaptive_avg_pool2d}
\alias{nnf_adaptive_avg_pool2d}
\title{Nnf adaptive avg pool2d}
\usage{
nnf_adaptive_avg_pool2d(self, output_size)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{output_size}{Parameter passed to the underlying ATen operator.}
}
\description{
Nnf adaptive avg pool2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
