% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_celu_}
\alias{nnf_celu_}
\title{Nnf celu}
\usage{
nnf_celu_(self, alpha)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{alpha}{Numeric scalar multiplier.}
}
\description{
Nnf celu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
