% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_conv_tbc}
\alias{nnf_conv_tbc}
\title{Nnf conv tbc}
\usage{
nnf_conv_tbc(self, weight, bias, pad)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{weight}{A `torch_tensor` of weights.}

\item{bias}{A `torch_tensor` of biases, or `NULL`.}

\item{pad}{Parameter passed to the underlying ATen operator.}
}
\description{
Nnf conv tbc
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
