% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_cosine_embedding_loss}
\alias{nnf_cosine_embedding_loss}
\title{Nnf cosine embedding loss}
\usage{
nnf_cosine_embedding_loss(input1, input2, target, margin, reduction)
}
\arguments{
\item{input1}{Parameter passed to the underlying ATen operator.}

\item{input2}{Parameter passed to the underlying ATen operator.}

\item{target}{Parameter passed to the underlying ATen operator.}

\item{margin}{Parameter passed to the underlying ATen operator.}

\item{reduction}{Parameter passed to the underlying ATen operator.}
}
\description{
Nnf cosine embedding loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
