% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_elu_}
\alias{nnf_elu_}
\title{Nnf elu}
\usage{
nnf_elu_(self, alpha, scale, input_scale)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{alpha}{Numeric scalar multiplier.}

\item{scale}{Parameter passed to the underlying ATen operator.}

\item{input_scale}{Parameter passed to the underlying ATen operator.}
}
\description{
Nnf elu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
