% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_embedding}
\alias{nnf_embedding}
\title{Nnf embedding}
\usage{
nnf_embedding(weight, indices)
}
\arguments{
\item{weight}{A `torch_tensor` of weights.}

\item{indices}{A `torch_tensor` of indices.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf embedding
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
