% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_glu}
\alias{nnf_glu}
\title{Nnf glu}
\usage{
nnf_glu(self, dim)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{dim}{Integer dimension.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf glu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
