% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_hardshrink}
\alias{nnf_hardshrink}
\title{Nnf hardshrink}
\usage{
nnf_hardshrink(self, lambd)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{lambd}{Parameter passed to the underlying ATen operator.}
}
\description{
Nnf hardshrink
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
