% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_hardsigmoid}
\alias{nnf_hardsigmoid}
\title{Nnf hardsigmoid}
\usage{
nnf_hardsigmoid(self)
}
\arguments{
\item{self}{A `torch_tensor`.}
}
\description{
Nnf hardsigmoid
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
