% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_hardswish}
\alias{nnf_hardswish}
\title{Nnf hardswish}
\usage{
nnf_hardswish(self)
}
\arguments{
\item{self}{A `torch_tensor`.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf hardswish
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
