% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_hardtanh}
\alias{nnf_hardtanh}
\title{Nnf hardtanh}
\usage{
nnf_hardtanh(self, min_val, max_val)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{min_val}{Parameter passed to the underlying ATen operator.}

\item{max_val}{Parameter passed to the underlying ATen operator.}
}
\description{
Nnf hardtanh
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
