% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_kl_div}
\alias{nnf_kl_div}
\title{Nnf kl div}
\usage{
nnf_kl_div(self, target, reduction, log_target)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{target}{Parameter passed to the underlying ATen operator.}

\item{reduction}{Parameter passed to the underlying ATen operator.}

\item{log_target}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf kl div
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
