% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_l1_loss}
\alias{nnf_l1_loss}
\title{Nnf l1 loss}
\usage{
nnf_l1_loss(self, target, reduction)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{target}{Parameter passed to the underlying ATen operator.}

\item{reduction}{Reduction string: `"none"`, `"mean"`, or `"sum"`.}
}
\description{
Nnf l1 loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
