% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_logsigmoid}
\alias{nnf_logsigmoid}
\title{Nnf logsigmoid}
\usage{
nnf_logsigmoid(input)
}
\arguments{
\item{input}{A `torch_tensor`.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf logsigmoid
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
