% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_one_hot}
\alias{nnf_one_hot}
\title{Nnf one hot}
\usage{
nnf_one_hot(self, num_classes)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{num_classes}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf one hot
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
