% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_pairwise_distance}
\alias{nnf_pairwise_distance}
\title{Nnf pairwise distance}
\usage{
nnf_pairwise_distance(x1, x2, p, eps, keepdim)
}
\arguments{
\item{x1}{Parameter passed to the underlying ATen operator.}

\item{x2}{Parameter passed to the underlying ATen operator.}

\item{p}{Parameter passed to the underlying ATen operator.}

\item{eps}{Numeric; small constant for numerical stability.}

\item{keepdim}{Logical; keep reduced dimensions.}
}
\description{
Nnf pairwise distance
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
