% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_pdist}
\alias{nnf_pdist}
\title{Nnf pdist}
\usage{
nnf_pdist(self, p)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{p}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf pdist
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
