% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_prelu}
\alias{nnf_prelu}
\title{Nnf prelu}
\usage{
nnf_prelu(self, weight)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{weight}{A `torch_tensor` of weights.}
}
\description{
Nnf prelu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
