% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_relu6}
\alias{nnf_relu6}
\title{Nnf relu6}
\usage{
nnf_relu6(self)
}
\arguments{
\item{self}{A `torch_tensor`.}
}
\description{
Nnf relu6
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
