% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_relu_}
\alias{nnf_relu_}
\title{Nnf relu}
\usage{
nnf_relu_(self)
}
\arguments{
\item{self}{A `torch_tensor`.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf relu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
