% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_rrelu_}
\alias{nnf_rrelu_}
\title{Nnf rrelu}
\usage{
nnf_rrelu_(self, lower, upper, training, generator)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{lower}{Parameter passed to the underlying ATen operator.}

\item{upper}{Parameter passed to the underlying ATen operator.}

\item{training}{Logical.}

\item{generator}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf rrelu
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
