% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_softmin}
\alias{nnf_softmin}
\title{Nnf softmin}
\usage{
nnf_softmin(input, dim = -1L)
}
\arguments{
\item{input}{A `torch_tensor`.}

\item{dim}{Integer dimension.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf softmin
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
