% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_threshold}
\alias{nnf_threshold}
\title{Nnf threshold}
\usage{
nnf_threshold(self, threshold, value)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{threshold}{Parameter passed to the underlying ATen operator.}

\item{value}{Parameter passed to the underlying ATen operator.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf threshold
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
