% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{optim_required}
\alias{optim_required}
\title{Required optimizer parameter sentinel}
\format{
An object of class \code{optim_required} of length 0.
}
\usage{
optim_required
}
\description{
Required optimizer parameter sentinel
}
\keyword{datasets}
