% tinyrox says don't edit this manually, but it can't stop you!
\name{optimizer}
\alias{optimizer}
\title{Optimizer}
\usage{
optimizer(name = NULL, ...)
}
\arguments{
\item{name}{Parameter passed to the underlying ATen operator.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Optimizer
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
