% tinyrox says don't edit this manually, but it can't stop you!
\name{print.torch_dtype}
\alias{print.torch_dtype}
\title{Print a torch_dtype}
\usage{
\method{print}{torch_dtype}(x, ...)
}
\arguments{
\item{x}{A torch_dtype.}

\item{...}{Ignored.}
}
\description{
Print a torch_dtype
}
\examples{
print(torch_float32)
}
