% tinyrox says don't edit this manually, but it can't stop you!
\name{print_segments}
\alias{print_segments}
\title{Print Segment Summary}
\usage{
print_segments(expr)
}
\arguments{
\item{expr}{An R expression}
}
\value{
Invisibly returns the segments
}
\description{
Pretty-prints how an expression would be segmented.
}
