% tinyrox says don't edit this manually, but it can't stop you!
\name{sampler}
\alias{sampler}
\title{Create a data sampler}
\usage{
sampler(data_source)
}
\arguments{
\item{data_source}{A dataset (used for length).}
}
\value{
A sampler object.
}
\description{
Create a data sampler
}
\examples{
\donttest{
if (torch_is_installed()) {
# Base sampler class; see dataloader() for usage
}
}
}
