% tinyrox says don't edit this manually, but it can't stop you!
\name{slc}
\alias{slc}
\title{Slc}
\usage{
slc(start = NULL, end = NULL, step = 1L)
}
\arguments{
\item{start}{Integer start index.}

\item{end}{Parameter passed to the underlying ATen operator.}

\item{step}{Parameter passed to the underlying ATen operator.}
}
\value{
A slice specification list.
}
\description{
Slc
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
