% tinyrox says don't edit this manually, but it can't stop you!
\name{torch___and__}
\alias{torch___and__}
\title{and}
\usage{
torch___and__(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::__and__}.
}
