% tinyrox says don't edit this manually, but it can't stop you!
\name{torch___ior__}
\alias{torch___ior__}
\title{ior}
\usage{
torch___ior__(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::__ior__}.
}
