% tinyrox says don't edit this manually, but it can't stop you!
\name{torch___irshift__}
\alias{torch___irshift__}
\title{irshift}
\usage{
torch___irshift__(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::__irshift__}.
}
