% tinyrox says don't edit this manually, but it can't stop you!
\name{torch___or__}
\alias{torch___or__}
\title{or}
\usage{
torch___or__(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::__or__}.
}
