% tinyrox says don't edit this manually, but it can't stop you!
\name{torch___rshift__}
\alias{torch___rshift__}
\title{rshift}
\usage{
torch___rshift__(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::__rshift__}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
