% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_absolute}
\alias{torch_absolute}
\title{Absolute}
\usage{
torch_absolute(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::absolute}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
