% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_acosh_}
\alias{torch_acosh_}
\title{Acosh}
\usage{
torch_acosh_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::acosh_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
