% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_adaptive_avg_pool1d}
\alias{torch_adaptive_avg_pool1d}
\title{Adaptive avg pool1d}
\usage{
torch_adaptive_avg_pool1d(self, output_size)
}
\arguments{
\item{self}{A torch_tensor.}

\item{output_size}{(int[1]).}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::adaptive_avg_pool1d}.
}
