% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_add}
\alias{torch_add}
\title{Add two tensors}
\usage{
torch_add(self, other, alpha = 1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}

\item{alpha}{Scalar multiplier for other (default 1).}
}
\description{
Add two tensors
}
