% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_addcdiv_}
\alias{torch_addcdiv_}
\title{Addcdiv}
\usage{
torch_addcdiv_(self, tensor1, tensor2, value = 1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{tensor1}{A torch_tensor.}

\item{tensor2}{A torch_tensor.}

\item{value}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::addcdiv_}.
}
