% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_align_as}
\alias{torch_align_as}
\title{Align as}
\usage{
torch_align_as(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::align_as}.
}
