% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_arctan2_}
\alias{torch_arctan2_}
\title{Arctan2}
\usage{
torch_arctan2_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::arctan2_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
