% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_arctanh}
\alias{torch_arctanh}
\title{Arctanh}
\usage{
torch_arctanh(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::arctanh}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
