% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_asin_}
\alias{torch_asin_}
\title{Asin}
\usage{
torch_asin_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::asin_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
