% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_asinh_}
\alias{torch_asinh_}
\title{Asinh}
\usage{
torch_asinh_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::asinh_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
