% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_atleast_3d}
\alias{torch_atleast_3d}
\title{Atleast 3d}
\usage{
torch_atleast_3d(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::atleast_3d}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
