% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bitwise_and_}
\alias{torch_bitwise_and_}
\title{Bitwise and}
\usage{
torch_bitwise_and_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bitwise_and_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
