% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bitwise_left_shift_}
\alias{torch_bitwise_left_shift_}
\title{Bitwise left shift}
\usage{
torch_bitwise_left_shift_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bitwise_left_shift_}.
}
