% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bitwise_not_}
\alias{torch_bitwise_not_}
\title{Bitwise not}
\usage{
torch_bitwise_not_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bitwise_not_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
