% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bitwise_or}
\alias{torch_bitwise_or}
\title{Bitwise or}
\usage{
torch_bitwise_or(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bitwise_or}.
}
