% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bitwise_right_shift}
\alias{torch_bitwise_right_shift}
\title{Bitwise right shift}
\usage{
torch_bitwise_right_shift(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bitwise_right_shift}.
}
