% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bitwise_xor}
\alias{torch_bitwise_xor}
\title{Bitwise xor}
\usage{
torch_bitwise_xor(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bitwise_xor}.
}
