% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_broadcast_tensors}
\alias{torch_broadcast_tensors}
\title{Broadcast tensors}
\usage{
torch_broadcast_tensors(tensors)
}
\arguments{
\item{tensors}{A list of torch_tensors.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::broadcast_tensors}.
}
