% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_broadcast_to}
\alias{torch_broadcast_to}
\title{Broadcast to}
\usage{
torch_broadcast_to(self, size)
}
\arguments{
\item{self}{A torch_tensor.}

\item{size}{An integer vector.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::broadcast_to}.
}
