% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_ccol_indices}
\alias{torch_ccol_indices}
\title{Ccol indices}
\usage{
torch_ccol_indices(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::ccol_indices}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
