% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_celu_}
\alias{torch_celu_}
\title{Celu}
\usage{
torch_celu_(self, alpha = 1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{alpha}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::celu_}.
}
