% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_chalf}
\alias{torch_chalf}
\title{Chalf}
\usage{
torch_chalf(self, memory_format = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{memory_format}{A memory format. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::chalf}.
}
