% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_cholesky}
\alias{torch_cholesky}
\title{Cholesky}
\usage{
torch_cholesky(self, upper = FALSE)
}
\arguments{
\item{self}{A torch_tensor.}

\item{upper}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::cholesky}.
}
