% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_clamp_max}
\alias{torch_clamp_max}
\title{Clamp max}
\usage{
torch_clamp_max(self, max)
}
\arguments{
\item{self}{A torch_tensor.}

\item{max}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::clamp_max}.
}
