% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_clone}
\alias{torch_clone}
\title{torch_clone}
\usage{
torch_clone(self)
}
\arguments{
\item{self}{A `torch_tensor`.}
}
\description{
torch_clone
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
