% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_conj_physical}
\alias{torch_conj_physical}
\title{Conj physical}
\usage{
torch_conj_physical(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::conj_physical}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
