% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_conv_transpose1d}
\alias{torch_conv_transpose1d}
\title{torch_conv_transpose1d}
\usage{
torch_conv_transpose1d(input, weight, bias = NULL, stride = 1L, padding = 0L,
                       output_padding = 0L, groups = 1L, dilation = 1L)
}
\description{
torch_conv_transpose1d
}
