% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_cumulative_trapezoid}
\alias{torch_cumulative_trapezoid}
\title{Cumulative trapezoid}
\usage{
torch_cumulative_trapezoid(y, x, dim = -1)
}
\arguments{
\item{y}{A torch_tensor.}

\item{x}{A torch_tensor.}

\item{dim}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::cumulative_trapezoid}.
}
