% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_dist}
\alias{torch_dist}
\title{Dist}
\usage{
torch_dist(self, other, p = 2)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}

\item{p}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::dist}.
}
