% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_divide}
\alias{torch_divide}
\title{Divide}
\usage{
torch_divide(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::divide}.
}
