% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_dtype_constants}
\alias{torch_dtype_constants}
\alias{torch_uint8}
\alias{torch_float16}
\alias{torch_float}
\alias{torch_float32}
\alias{torch_float64}
\alias{torch_double}
\alias{torch_int8}
\alias{torch_int16}
\alias{torch_int}
\alias{torch_int32}
\alias{torch_int64}
\alias{torch_long}
\alias{torch_bool}
\title{Torch Data Types}
\description{
Constants representing tensor element types, matching libtorch's
\code{c10::ScalarType} enum. Pass these to \code{dtype} arguments
in tensor creation functions.
}
\examples{
torch_float32
torch_int64
}
