% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_erf_}
\alias{torch_erf_}
\title{Erf}
\usage{
torch_erf_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::erf_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
