% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_expand_as}
\alias{torch_expand_as}
\title{Expand as}
\usage{
torch_expand_as(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::expand_as}.
}
