% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_flatten_dense_tensors}
\alias{torch_flatten_dense_tensors}
\title{Flatten dense tensors}
\usage{
torch_flatten_dense_tensors(tensors)
}
\arguments{
\item{tensors}{A list of torch_tensors.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::flatten_dense_tensors}.
}
