% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_floor_divide_}
\alias{torch_floor_divide_}
\title{Floor divide}
\usage{
torch_floor_divide_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::floor_divide_}.
}
