% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_full}
\alias{torch_full}
\title{Create a tensor filled with a scalar value}
\usage{
torch_full(size, fill_value, dtype = NULL, device = NULL)
}
\arguments{
\item{size}{Integer vector of dimensions.}

\item{fill_value}{Scalar fill value.}

\item{dtype}{Optional torch dtype.}

\item{device}{Ignored (CPU only).}
}
\value{
A torch_tensor.
}
\description{
Create a tensor filled with a scalar value
}
