% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_geometric_}
\alias{torch_geometric_}
\title{Geometric}
\usage{
torch_geometric_(self, p, generator = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{p}{A numeric value.}

\item{generator}{A random number generator. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::geometric_}.
}
