% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_get_default_dtype}
\alias{torch_get_default_dtype}
\alias{torch_set_default_dtype}
\title{Get/set default dtype}
\usage{
torch_get_default_dtype()

torch_set_default_dtype(dtype)
}
\arguments{
\item{dtype}{A torch_dtype to set as default.}
}
\value{
For get, the current default dtype. For set, invisible NULL.
}
\description{
Get/set default dtype
}
