% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_get_num_interop_threads}
\alias{torch_get_num_interop_threads}
\title{Get Number of Interop Threads}
\usage{
torch_get_num_interop_threads()
}
\value{
Integer number of interop threads.
}
\description{
Get Number of Interop Threads
}
