% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_greater}
\alias{torch_greater}
\title{Greater}
\usage{
torch_greater(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::greater}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
