% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_greater_equal}
\alias{torch_greater_equal}
\title{Greater equal}
\usage{
torch_greater_equal(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::greater_equal}.
}
