% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_heaviside}
\alias{torch_heaviside}
\title{Heaviside}
\usage{
torch_heaviside(self, values)
}
\arguments{
\item{self}{A torch_tensor.}

\item{values}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::heaviside}.
}
