% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_igamma}
\alias{torch_igamma}
\title{Igamma}
\usage{
torch_igamma(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::igamma}.
}
