% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_igammac}
\alias{torch_igammac}
\title{Igammac}
\usage{
torch_igammac(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::igammac}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
