% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_index}
\alias{torch_index}
\title{torch_index}
\usage{
torch_index(self, indices)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{indices}{A `torch_tensor` of indices.}
}
\description{
torch_index
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
