% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_index_put}
\alias{torch_index_put}
\title{torch_index_put}
\usage{
torch_index_put(self, indices, values)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{indices}{A `torch_tensor` of indices.}

\item{values}{A `torch_tensor` of values.}
}
\description{
torch_index_put
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
