% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_index_select}
\alias{torch_index_select}
\title{torch_index_select}
\usage{
torch_index_select(self, dim, index)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{dim}{Integer dimension.}

\item{index}{A `torch_tensor` of indices.}
}
\value{
A `torch_tensor`.
}
\description{
torch_index_select
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
