% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_is_complex}
\alias{torch_is_complex}
\title{Dtype query: is the tensor complex?}
\usage{
torch_is_complex(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
Logical.
}
\description{
Dtype query: is the tensor complex?
}
