% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_is_floating_point}
\alias{torch_is_floating_point}
\title{Dtype query: is the tensor floating point?}
\usage{
torch_is_floating_point(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
Logical.
}
\description{
Dtype query: is the tensor floating point?
}
