% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_is_installed}
\alias{torch_is_installed}
\alias{torch_install_path}
\title{Check if tinytorch backend is installed}
\usage{
torch_is_installed()

torch_install_path()
}
\value{
Logical.
}
\description{
Check if tinytorch backend is installed
}
