% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_lcm_}
\alias{torch_lcm_}
\title{Lcm}
\usage{
torch_lcm_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::lcm_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
