% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_ldexp_}
\alias{torch_ldexp_}
\title{Ldexp}
\usage{
torch_ldexp_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::ldexp_}.
}
