% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_less_equal_}
\alias{torch_less_equal_}
\title{Less equal}
\usage{
torch_less_equal_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::less_equal_}.
}
