% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_cond}
\alias{torch_linalg_cond}
\title{Linalg cond}
\usage{
torch_linalg_cond(self, p = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{p}{A numeric scalar. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_cond}.
}
