% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_det}
\alias{torch_linalg_det}
\alias{linalg_det}
\title{Linalg det}
\usage{
torch_linalg_det(A)
}
\arguments{
\item{A}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_det}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
