% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_diagonal}
\alias{torch_linalg_diagonal}
\title{Linalg diagonal}
\usage{
torch_linalg_diagonal(A, offset = 0, dim1 = -2, dim2 = -1)
}
\arguments{
\item{A}{A torch_tensor.}

\item{offset}{An integer.}

\item{dim1}{An integer.}

\item{dim2}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_diagonal}.
}
