% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_inv_ex}
\alias{torch_linalg_inv_ex}
\title{Linalg inv ex}
\usage{
torch_linalg_inv_ex(A, check_errors = FALSE)
}
\arguments{
\item{A}{A torch_tensor.}

\item{check_errors}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_inv_ex}.
}
