% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_ldl_factor_ex}
\alias{torch_linalg_ldl_factor_ex}
\title{Linalg ldl factor ex}
\usage{
torch_linalg_ldl_factor_ex(self, hermitian = FALSE, check_errors = FALSE)
}
\arguments{
\item{self}{A torch_tensor.}

\item{hermitian}{A logical value.}

\item{check_errors}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_ldl_factor_ex}.
}
