% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_lu_factor_ex}
\alias{torch_linalg_lu_factor_ex}
\title{Linalg lu factor ex}
\usage{
torch_linalg_lu_factor_ex(A, pivot = TRUE, check_errors = FALSE)
}
\arguments{
\item{A}{A torch_tensor.}

\item{pivot}{A logical value.}

\item{check_errors}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_lu_factor_ex}.
}
