% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_matmul}
\alias{torch_linalg_matmul}
\title{Linalg matmul}
\usage{
torch_linalg_matmul(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_matmul}.
}
