% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_solve_ex}
\alias{torch_linalg_solve_ex}
\title{Linalg solve ex}
\usage{
torch_linalg_solve_ex(A, B, left = TRUE, check_errors = FALSE)
}
\arguments{
\item{A}{A torch_tensor.}

\item{B}{A torch_tensor.}

\item{left}{A logical value.}

\item{check_errors}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_solve_ex}.
}
