% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_tensorsolve}
\alias{torch_linalg_tensorsolve}
\title{Linalg tensorsolve}
\usage{
torch_linalg_tensorsolve(self, other, dims = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}

\item{dims}{An integer vector. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_tensorsolve}.
}
