% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_vecdot}
\alias{torch_linalg_vecdot}
\title{Linalg vecdot}
\usage{
torch_linalg_vecdot(x, y, dim = -1)
}
\arguments{
\item{x}{A torch_tensor.}

\item{y}{A torch_tensor.}

\item{dim}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_vecdot}.
}
