% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linear}
\alias{torch_linear}
\title{Linear transformation}
\usage{
torch_linear(input, weight, bias = NULL)
}
\arguments{
\item{input}{Input tensor.}

\item{weight}{Weight matrix.}

\item{bias}{Optional bias vector.}
}
\description{
Linear transformation
}
