% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_log_sigmoid_forward}
\alias{torch_log_sigmoid_forward}
\title{Log sigmoid forward}
\usage{
torch_log_sigmoid_forward(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::log_sigmoid_forward}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
