% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_log_softmax}
\alias{torch_log_softmax}
\title{torch_log_softmax}
\usage{
torch_log_softmax(self, dim = -1L)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{dim}{Integer dimension.}
}
\description{
torch_log_softmax
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
