% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_logaddexp2}
\alias{torch_logaddexp2}
\title{Logaddexp2}
\usage{
torch_logaddexp2(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::logaddexp2}.
}
