% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_logical_and_}
\alias{torch_logical_and_}
\title{Logical and}
\usage{
torch_logical_and_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::logical_and_}.
}
