% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_logical_or_}
\alias{torch_logical_or_}
\title{Logical or}
\usage{
torch_logical_or_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::logical_or_}.
}
