% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_logit_}
\alias{torch_logit_}
\title{Logit}
\usage{
torch_logit_(self, eps = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{eps}{A numeric value. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::logit_}.
}
