% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_lt_}
\alias{torch_lt_}
\title{Lt}
\usage{
torch_lt_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::lt_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
