% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_manual_seed}
\alias{torch_manual_seed}
\title{Set the random seed for torch operations}
\usage{
torch_manual_seed(seed)
}
\arguments{
\item{seed}{Integer seed value.}
}
\value{
Invisible NULL.
}
\description{
Set the random seed for torch operations
}
