% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_masked_scatter_}
\alias{torch_masked_scatter_}
\title{Masked scatter}
\usage{
torch_masked_scatter_(self, mask, source)
}
\arguments{
\item{self}{A torch_tensor.}

\item{mask}{A torch_tensor.}

\item{source}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::masked_scatter_}.
}
