% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_matrix_power}
\alias{torch_matrix_power}
\title{Matrix power}
\usage{
torch_matrix_power(self, n)
}
\arguments{
\item{self}{A torch_tensor.}

\item{n}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::matrix_power}.
}
