% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_mode}
\alias{torch_mode}
\title{Mode}
\usage{
torch_mode(self, dim = -1, keepdim = FALSE)
}
\arguments{
\item{self}{A torch_tensor.}

\item{dim}{An integer.}

\item{keepdim}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::mode}.
}
