% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_moveaxis}
\alias{torch_moveaxis}
\title{Moveaxis}
\usage{
torch_moveaxis(self, source, destination)
}
\arguments{
\item{self}{A torch_tensor.}

\item{source}{An integer vector.}

\item{destination}{An integer vector.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::moveaxis}.
}
