% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_mul}
\alias{torch_mul}
\title{Multiply two tensors element-wise}
\usage{
torch_mul(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor or scalar.}
}
\value{
A `torch_tensor`.
}
\description{
Multiply two tensors element-wise
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
