% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_mvlgamma_}
\alias{torch_mvlgamma_}
\title{Mvlgamma}
\usage{
torch_mvlgamma_(self, p)
}
\arguments{
\item{self}{A torch_tensor.}

\item{p}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::mvlgamma_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
