% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_narrow}
\alias{torch_narrow}
\title{torch_narrow}
\usage{
torch_narrow(self, dim, start, length)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{dim}{Integer dimension.}

\item{start}{Integer start index.}

\item{length}{Integer length.}
}
\value{
A `torch_tensor`.
}
\description{
torch_narrow
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
