% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_nonzero_static}
\alias{torch_nonzero_static}
\title{Nonzero static}
\usage{
torch_nonzero_static(self, size, fill_value = -1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{size}{An integer.}

\item{fill_value}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::nonzero_static}.
}
