% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_not_equal}
\alias{torch_not_equal}
\title{Not equal}
\usage{
torch_not_equal(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::not_equal}.
}
