% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_numpy_T}
\alias{torch_numpy_T}
\title{Numpy T}
\usage{
torch_numpy_T(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::numpy_T}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
