% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_one_hot}
\alias{torch_one_hot}
\title{One hot}
\usage{
torch_one_hot(self, num_classes = -1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{num_classes}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::one_hot}.
}
