% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_pinverse}
\alias{torch_pinverse}
\title{Pinverse}
\usage{
torch_pinverse(self, rcond = 1e-15)
}
\arguments{
\item{self}{A torch_tensor.}

\item{rcond}{A numeric value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::pinverse}.
}
