% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_poisson}
\alias{torch_poisson}
\title{Poisson}
\usage{
torch_poisson(self, generator = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{generator}{A random number generator. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::poisson}.
}
