% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_polygamma_}
\alias{torch_polygamma_}
\title{Polygamma}
\usage{
torch_polygamma_(self, n)
}
\arguments{
\item{self}{A torch_tensor.}

\item{n}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::polygamma_}.
}
