% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_positive}
\alias{torch_positive}
\title{Positive}
\usage{
torch_positive(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::positive}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
