% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_pow}
\alias{torch_pow}
\title{Power}
\usage{
torch_pow(self, exponent)
}
\arguments{
\item{self}{A torch_tensor or numeric scalar.}

\item{exponent}{A torch_tensor or numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Power
}
