% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_q_per_channel_scales}
\alias{torch_q_per_channel_scales}
\title{Q per channel scales}
\usage{
torch_q_per_channel_scales(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::q_per_channel_scales}.
}
