% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_q_per_channel_zero_points}
\alias{torch_q_per_channel_zero_points}
\title{Q per channel zero points}
\usage{
torch_q_per_channel_zero_points(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::q_per_channel_zero_points}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
