% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_q_scale}
\alias{torch_q_scale}
\title{Q scale}
\usage{
torch_q_scale(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::q_scale}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
