% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_rad2deg_}
\alias{torch_rad2deg_}
\title{Rad2deg}
\usage{
torch_rad2deg_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::rad2deg_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
