% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_randn_like}
\alias{torch_randn_like}
\title{Create a tensor of random normal values with same shape/dtype as input}
\usage{
torch_randn_like(input, dtype = NULL)
}
\arguments{
\item{dtype}{Optional dtype override.}

\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Create a tensor of random normal values with same shape/dtype as input
}
