% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_reflection_pad2d}
\alias{torch_reflection_pad2d}
\title{Reflection pad2d}
\usage{
torch_reflection_pad2d(self, padding)
}
\arguments{
\item{self}{A torch_tensor.}

\item{padding}{(int[4]).}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::reflection_pad2d}.
}
