% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_replication_pad1d}
\alias{torch_replication_pad1d}
\title{Replication pad1d}
\usage{
torch_replication_pad1d(self, padding)
}
\arguments{
\item{self}{A torch_tensor.}

\item{padding}{(int[2]).}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::replication_pad1d}.
}
