% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_replication_pad3d}
\alias{torch_replication_pad3d}
\title{Replication pad3d}
\usage{
torch_replication_pad3d(self, padding)
}
\arguments{
\item{self}{A torch_tensor.}

\item{padding}{(int[6]).}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::replication_pad3d}.
}
