% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_reshape_as}
\alias{torch_reshape_as}
\title{Reshape as}
\usage{
torch_reshape_as(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::reshape_as}.
}
