% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_resize_as_}
\alias{torch_resize_as_}
\title{Resize as}
\usage{
torch_resize_as_(self, the_template, memory_format = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{the_template}{A torch_tensor.}

\item{memory_format}{A memory format. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::resize_as_}.
}
