% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_resolve_conj}
\alias{torch_resolve_conj}
\title{Resolve conj}
\usage{
torch_resolve_conj(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::resolve_conj}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
