% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_result_type}
\alias{torch_result_type}
\title{Result type}
\usage{
torch_result_type(tensor1, other)
}
\arguments{
\item{other}{A torch_tensor.}

\item{tensor}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::result_type}.
}
