% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_selu}
\alias{torch_selu}
\title{Selu}
\usage{
torch_selu(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::selu}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
