% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_sparse_dim}
\alias{torch_sparse_dim}
\title{Sparse dim}
\usage{
torch_sparse_dim(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::sparse_dim}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
