% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_airy_ai}
\alias{torch_special_airy_ai}
\title{Special airy ai}
\usage{
torch_special_airy_ai(x)
}
\arguments{
\item{x}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_airy_ai}.
}
