% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_chebyshev_polynomial_v}
\alias{torch_special_chebyshev_polynomial_v}
\title{Special chebyshev polynomial v}
\usage{
torch_special_chebyshev_polynomial_v(x, n)
}
\arguments{
\item{x}{A torch_tensor.}

\item{n}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_chebyshev_polynomial_v}.
}
