% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_digamma}
\alias{torch_special_digamma}
\title{Special digamma}
\usage{
torch_special_digamma(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_digamma}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
