% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_erfcx}
\alias{torch_special_erfcx}
\title{Special erfcx}
\usage{
torch_special_erfcx(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_erfcx}.
}
