% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_expm1}
\alias{torch_special_expm1}
\title{Special expm1}
\usage{
torch_special_expm1(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_expm1}.
}
