% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_gammaincc}
\alias{torch_special_gammaincc}
\title{Special gammaincc}
\usage{
torch_special_gammaincc(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_gammaincc}.
}
