% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_gammaln}
\alias{torch_special_gammaln}
\title{Special gammaln}
\usage{
torch_special_gammaln(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_gammaln}.
}
