% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_laguerre_polynomial_l}
\alias{torch_special_laguerre_polynomial_l}
\title{Special laguerre polynomial l}
\usage{
torch_special_laguerre_polynomial_l(x, n)
}
\arguments{
\item{x}{A torch_tensor.}

\item{n}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_laguerre_polynomial_l}.
}
