% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_legendre_polynomial_p}
\alias{torch_special_legendre_polynomial_p}
\title{Special legendre polynomial p}
\usage{
torch_special_legendre_polynomial_p(x, n)
}
\arguments{
\item{x}{A torch_tensor.}

\item{n}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_legendre_polynomial_p}.
}
