% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_log_ndtr}
\alias{torch_special_log_ndtr}
\title{Special log ndtr}
\usage{
torch_special_log_ndtr(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_log_ndtr}.
}
