% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_logit}
\alias{torch_special_logit}
\title{Special logit}
\usage{
torch_special_logit(self, eps = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{eps}{A numeric value. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_logit}.
}
