% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_modified_bessel_i0}
\alias{torch_special_modified_bessel_i0}
\title{Special modified bessel i0}
\usage{
torch_special_modified_bessel_i0(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_modified_bessel_i0}.
}
