% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_ndtri}
\alias{torch_special_ndtri}
\title{Special ndtri}
\usage{
torch_special_ndtri(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_ndtri}.
}
