% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_psi}
\alias{torch_special_psi}
\title{Special psi}
\usage{
torch_special_psi(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_psi}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
