% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_scaled_modified_bessel_k0}
\alias{torch_special_scaled_modified_bessel_k0}
\title{Special scaled modified bessel k0}
\usage{
torch_special_scaled_modified_bessel_k0(x)
}
\arguments{
\item{x}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_scaled_modified_bessel_k0}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
