% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_scaled_modified_bessel_k1}
\alias{torch_special_scaled_modified_bessel_k1}
\title{Special scaled modified bessel k1}
\usage{
torch_special_scaled_modified_bessel_k1(x)
}
\arguments{
\item{x}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_scaled_modified_bessel_k1}.
}
