% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_shifted_chebyshev_polynomial_t}
\alias{torch_special_shifted_chebyshev_polynomial_t}
\title{Special shifted chebyshev polynomial t}
\usage{
torch_special_shifted_chebyshev_polynomial_t(x, n)
}
\arguments{
\item{x}{A torch_tensor.}

\item{n}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_shifted_chebyshev_polynomial_t}.
}
