% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_shifted_chebyshev_polynomial_u}
\alias{torch_special_shifted_chebyshev_polynomial_u}
\title{Special shifted chebyshev polynomial u}
\usage{
torch_special_shifted_chebyshev_polynomial_u(x, n)
}
\arguments{
\item{x}{A torch_tensor.}

\item{n}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_shifted_chebyshev_polynomial_u}.
}
