% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_shifted_chebyshev_polynomial_w}
\alias{torch_special_shifted_chebyshev_polynomial_w}
\title{Special shifted chebyshev polynomial w}
\usage{
torch_special_shifted_chebyshev_polynomial_w(x, n)
}
\arguments{
\item{x}{A torch_tensor.}

\item{n}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_shifted_chebyshev_polynomial_w}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
