% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_sqrt_}
\alias{torch_sqrt_}
\title{Sqrt}
\usage{
torch_sqrt_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::sqrt_}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
