% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_subtract}
\alias{torch_subtract}
\title{Subtract}
\usage{
torch_subtract(self, other, alpha = 1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}

\item{alpha}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::subtract}.
}
