% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_swapaxes}
\alias{torch_swapaxes}
\title{Swapaxes}
\usage{
torch_swapaxes(self, axis0, axis1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{axis0}{An integer.}

\item{axis1}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::swapaxes}.
}
