% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_swapdims}
\alias{torch_swapdims}
\title{Swapdims}
\usage{
torch_swapdims(self, dim0, dim1)
}
\arguments{
\item{self}{A torch_tensor.}

\item{dim0}{An integer.}

\item{dim1}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::swapdims}.
}
