% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_to_sparse}
\alias{torch_to_sparse}
\title{To sparse}
\usage{
torch_to_sparse(self, sparse_dim)
}
\arguments{
\item{self}{A torch_tensor.}

\item{sparse_dim}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::to_sparse}.
}
