% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_triu_}
\alias{torch_triu_}
\title{Triu}
\usage{
torch_triu_(self, diagonal = 0)
}
\arguments{
\item{self}{A torch_tensor.}

\item{diagonal}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::triu_}.
}
