% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_true_divide_}
\alias{torch_true_divide_}
\title{True divide}
\usage{
torch_true_divide_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::true_divide_}.
}
