% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_type_as}
\alias{torch_type_as}
\title{Type as}
\usage{
torch_type_as(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::type_as}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
