% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_var_mean}
\alias{torch_var_mean}
\title{Var mean}
\usage{
torch_var_mean(self, unbiased = TRUE)
}
\arguments{
\item{self}{A torch_tensor.}

\item{unbiased}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::var_mean}.
}
