% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_view_as}
\alias{torch_view_as}
\title{View as}
\usage{
torch_view_as(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::view_as}.
}
