% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_xlogy_}
\alias{torch_xlogy_}
\title{Xlogy}
\usage{
torch_xlogy_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::xlogy_}.
}
