% tinyrox says don't edit this manually, but it can't stop you!
\name{with_detect_anomaly}
\alias{with_detect_anomaly}
\title{With detect anomaly}
\usage{
with_detect_anomaly(code)
}
\arguments{
\item{code}{Parameter passed to the underlying ATen operator.}
}
\value{
The result of evaluating `code`.
}
\description{
With detect anomaly
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
