% tinyrox says don't edit this manually, but it can't stop you!
\name{with_device}
\alias{with_device}
\title{With device}
\usage{
with_device(device, code)
}
\arguments{
\item{device}{Parameter passed to the underlying ATen operator.}

\item{code}{Parameter passed to the underlying ATen operator.}
}
\value{
The result of evaluating `code`.
}
\description{
With device
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
