% tinyrox says don't edit this manually, but it can't stop you!
\name{with_enable_grad}
\alias{with_enable_grad}
\title{With enable grad}
\usage{
with_enable_grad(code)
}
\arguments{
\item{code}{Parameter passed to the underlying ATen operator.}
}
\value{
The result of evaluating `code`.
}
\description{
With enable grad
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
