% tinyrox says don't edit this manually, but it can't stop you!
\name{yield}
\alias{yield}
\title{Yield placeholder (for compatibility)}
\usage{
yield(x)
}
\arguments{
\item{x}{Value to yield.}
}
\value{
x
}
\description{
Yield placeholder (for compatibility)
}
\examples{
\donttest{
if (torch_is_installed()) {
# Used inside iterable_dataset definitions; see dataset() docs
}
}
}
\keyword{internal}
