% tinyrox says don't edit this manually, but it can't stop you!
\name{cuda_empty_cache}
\alias{cuda_empty_cache}
\title{Release cached CUDA memory}
\usage{
cuda_empty_cache()
}
\description{
Releases all unoccupied cached memory held by the CUDA allocator
so that it can be used by other GPU applications.
}
