% tinyrox says don't edit this manually, but it can't stop you!
\name{distr_mixture_same_family}
\alias{distr_mixture_same_family}
\title{Mixture of same-family distributions}
\usage{
distr_mixture_same_family(mixture_distribution, component_distribution)
}
\arguments{
\item{mixture_distribution}{Categorical distribution for mixture weights.}

\item{component_distribution}{Batch of component distributions.}
}
\value{
A distribution object.
}
\description{
Mixture of same-family distributions
}
