% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_log_sigmoid}
\alias{nn_log_sigmoid}
\title{Nn log sigmoid}
\usage{
nn_log_sigmoid()
}
\value{
An `nn_module` instance.
}
\description{
Nn log sigmoid
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
