% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_log_softmax}
\alias{nn_log_softmax}
\title{Nn log softmax}
\usage{
nn_log_softmax(dim = -1L)
}
\arguments{
\item{dim}{Integer dimension.}
}
\description{
Nn log softmax
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
