% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_module_dict}
\alias{nn_module_dict}
\title{Module dictionary}
\usage{
nn_module_dict(modules = list())
}
\arguments{
\item{modules}{Named list of nn_module instances.}
}
\value{
An nn_module that stores sub-modules by name.
}
\description{
Module dictionary
}
