% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_multilabel_margin_loss}
\alias{nn_multilabel_margin_loss}
\title{Nn multilabel margin loss}
\usage{
nn_multilabel_margin_loss(reduction = 1L)
}
\arguments{
\item{reduction}{Parameter passed to the underlying ATen operator.}
}
\value{
An `nn_module` instance.
}
\description{
Nn multilabel margin loss
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
