% tinyrox says don't edit this manually, but it can't stop you!
\name{nn_tanhshrink}
\alias{nn_tanhshrink}
\title{Nn tanhshrink}
\usage{
nn_tanhshrink()
}
\description{
Nn tanhshrink
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
