% tinyrox says don't edit this manually, but it can't stop you!
\name{nnf_dropout2d}
\alias{nnf_dropout2d}
\title{Nnf dropout2d}
\usage{
nnf_dropout2d(input, p = 0.5, training = TRUE)
}
\arguments{
\item{input}{A `torch_tensor`.}

\item{p}{Numeric norm exponent.}

\item{training}{Logical.}
}
\value{
A `torch_tensor`.
}
\description{
Nnf dropout2d
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
