% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_atan2_}
\alias{torch_atan2_}
\title{Atan2}
\usage{
torch_atan2_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::atan2_}.
}
