% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_batch_norm}
\alias{torch_batch_norm}
\title{torch_batch_norm}
\usage{
torch_batch_norm(input, weight = NULL, bias = NULL, running_mean = NULL,
                 running_var = NULL, training = FALSE, momentum = 0.1,
                 eps = 1e-05, cudnn_enabled = TRUE)
}
\description{
torch_batch_norm
}
