% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_bernoulli}
\alias{torch_bernoulli}
\title{Bernoulli}
\usage{
torch_bernoulli(self, generator = NULL)
}
\arguments{
\item{self}{A torch_tensor.}

\item{generator}{A random number generator. Optional.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::bernoulli}.
}
