% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_block_diag}
\alias{torch_block_diag}
\title{Block diag}
\usage{
torch_block_diag(tensors)
}
\arguments{
\item{tensors}{A list of torch_tensors.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::block_diag}.
}
