% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_cartesian_prod}
\alias{torch_cartesian_prod}
\title{Cartesian prod}
\usage{
torch_cartesian_prod(tensors)
}
\arguments{
\item{tensors}{A list of torch_tensors.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::cartesian_prod}.
}
