% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_chain_matmul}
\alias{torch_chain_matmul}
\title{Chain matmul}
\usage{
torch_chain_matmul(matrices)
}
\arguments{
\item{matrices}{A list of torch_tensors.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::chain_matmul}.
}
