% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_conj_physical_}
\alias{torch_conj_physical_}
\title{Conj physical}
\usage{
torch_conj_physical_(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::conj_physical_}.
}
