% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_diagflat}
\alias{torch_diagflat}
\title{Diagflat}
\usage{
torch_diagflat(self, offset = 0)
}
\arguments{
\item{self}{A torch_tensor.}

\item{offset}{An integer.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::diagflat}.
}
