% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_divide_}
\alias{torch_divide_}
\title{Divide}
\usage{
torch_divide_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::divide_}.
}
