% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_floor_divide}
\alias{torch_floor_divide}
\title{torch_floor_divide}
\usage{
torch_floor_divide(self, other)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{other}{A `torch_tensor` or scalar.}
}
\description{
torch_floor_divide
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
