% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_generator}
\alias{torch_generator}
\title{Create a random number generator (stub)}
\usage{
torch_generator()
}
\value{
A generator object (currently an environment placeholder).
}
\description{
Create a random number generator (stub)
}
