% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_linalg_matrix_exp}
\alias{torch_linalg_matrix_exp}
\title{Linalg matrix exp}
\usage{
torch_linalg_matrix_exp(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::linalg_matrix_exp}.
}
