% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_logical_and}
\alias{torch_logical_and}
\title{Logical and}
\usage{
torch_logical_and(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::logical_and}.
}
