% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_min}
\alias{torch_min}
\title{torch_min}
\usage{
torch_min(self, dim = NULL)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{dim}{Integer dimension.}
}
\description{
torch_min
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
