% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_nonzero_numpy}
\alias{torch_nonzero_numpy}
\title{Nonzero numpy}
\usage{
torch_nonzero_numpy(self)
}
\arguments{
\item{self}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::nonzero_numpy}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
