% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_orgqr}
\alias{torch_orgqr}
\title{Orgqr}
\usage{
torch_orgqr(self, input2)
}
\arguments{
\item{self}{A torch_tensor.}

\item{input2}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::orgqr}.
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
