% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_outer}
\alias{torch_outer}
\title{Outer product of two vectors}
\usage{
torch_outer(self, vec2)
}
\arguments{
\item{self}{First vector.}

\item{vec2}{Second vector.}
}
\value{
A torch_tensor.
}
\description{
Outer product of two vectors
}
