% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_polar}
\alias{torch_polar}
\title{Create a complex tensor from polar coordinates}
\usage{
torch_polar(abs, angle)
}
\arguments{
\item{abs}{Absolute value tensor.}

\item{angle}{Angle tensor.}
}
\value{
A torch_tensor.
}
\description{
Create a complex tensor from polar coordinates
}
