% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_randn}
\alias{torch_randn}
\title{Create a tensor with random normal values}
\usage{
torch_randn(..., dtype = NULL, device = NULL)
}
\arguments{
\item{...}{Dimensions (as separate integers or a single integer vector).}

\item{dtype}{Optional torch dtype.}

\item{device}{Ignored (CPU only).}
}
\description{
Create a tensor with random normal values
}
