% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_remainder}
\alias{torch_remainder}
\title{torch_remainder}
\usage{
torch_remainder(self, other)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{other}{A `torch_tensor` or scalar.}
}
\value{
A `torch_tensor`.
}
\description{
torch_remainder
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
