% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_remainder_}
\alias{torch_remainder_}
\title{Remainder}
\usage{
torch_remainder_(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A numeric scalar.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::remainder_}.
}
