% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_repeat_interleave}
\alias{torch_repeat_interleave}
\title{torch_repeat_interleave}
\usage{
torch_repeat_interleave(self, repeats, dim = NULL)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{repeats}{Integer or `torch_tensor` of repeat counts.}

\item{dim}{Integer dimension.}
}
\value{
A `torch_tensor`.
}
\description{
torch_repeat_interleave
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
