% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_reshape}
\alias{torch_reshape}
\title{torch_reshape}
\usage{
torch_reshape(self, shape)
}
\arguments{
\item{self}{A `torch_tensor`.}

\item{shape}{Integer vector shape.}
}
\description{
torch_reshape
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
