% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_special_xlogy}
\alias{torch_special_xlogy}
\title{Special xlogy}
\usage{
torch_special_xlogy(self, other)
}
\arguments{
\item{self}{A torch_tensor.}

\item{other}{A torch_tensor.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::special_xlogy}.
}
