% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_svd}
\alias{torch_svd}
\title{Svd}
\usage{
torch_svd(self, some = TRUE, compute_uv = TRUE)
}
\arguments{
\item{self}{A torch_tensor.}

\item{some}{A logical value.}

\item{compute_uv}{A logical value.}
}
\value{
A torch_tensor.
}
\description{
Wrapper for \code{at::svd}.
}
