% tinyrox says don't edit this manually, but it can't stop you!
\name{torch_triu}
\alias{torch_triu}
\title{Upper triangular part of a matrix}
\usage{
torch_triu(self, diagonal = 0L)
}
\arguments{
\item{self}{A torch_tensor.}

\item{diagonal}{Offset from the main diagonal. Default 0.}
}
\value{
A torch_tensor.
}
\description{
Upper triangular part of a matrix
}
