% tinyrox says don't edit this manually, but it can't stop you!
\name{unset_autocast}
\alias{unset_autocast}
\title{Unset autocast}
\usage{
unset_autocast()
}
\description{
Unset autocast
}
\examples{
\donttest{
if (torch_is_installed()) {
  # See PyTorch docs: https://docs.pytorch.org/docs/stable/torch.html
}
}
}
